LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/navigate-same-document-intercept-reject_no-currententrychange-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: boo

Harness Error (FAIL), message = Unhandled rejection: boo

PASS event and promise ordering for same-document navigation.navigate() intercepted by passing a rejected promise to intercept()
